+++ Packs/SplunkPy_v2/CHANGELOG.md
@@ -1,1 +1,2 @@
 ## [Unreleased]
+- New SplunkPy v2 integration.